+++ examples/echobot.go
@@ -38,7 +38,7 @@ func main() {
 	})
 	cli.OnBotStart(func(bot *deltachat.Bot, cmd *cobra.Command, args []string) {
 		addr, _ := bot.GetConfig("addr")
-		cli.Logger.Info().Msgf("Listening at: %v", addr)
+		cli.Logger.Infof("Listening at: %v", addr)
 	})
 	cli.Start()
 }